chore: rename back to correct directory names
