+++ packages/datepickers/demo/datepicker.stories.mdx
@@ -1,7 +1,7 @@
 import { Meta, ArgsTable, Canvas, Story, Markdown } from '@storybook/addon-docs';
 import { useArgs } from '@storybook/client-api';
 import { DatePicker } from '@zendeskgarden/react-datepickers';
-import { DatePickerStory } from './stories/DatePickerStory';
+import { DatePickerStory } from './stories/Date_PickerStory';
 import { DATE_STYLE_OPTIONS } from './stories/data';
 import README from '../README.md';
