+++ packages/datepickers/src/elements/DatePicker/components/MonthSelector.tsx
@@ -7,7 +7,7 @@
 
 import React, { useCallback } from 'react';
 import { StyledHeader, StyledHeaderPaddle, StyledHeaderLabel } from '../../../styled';
-import useDatePickerContext from '../utils/useDatePickerContext';
+import useDatePickerContext from '../utils/useDate_PickerContext';
 
 import ChevronLeftStrokeIcon from '@zendeskgarden/svg-icons/src/16/chevron-left-stroke.svg';
 import ChevronRightStrokeIcon from '@zendeskgarden/svg-icons/src/16/chevron-right-stroke.svg';
