+++ packages/datepickers/src/elements/DatePickerRange/utils/date-picker-range-reducer.ts
